Move Index and Document to utils.py
